pyproject.toml
[tool.poetry]
name = "catlogbeacon"
version = "0.1.0"
description = ""
authors = ["Hi-king <hikingko1@gmail.com>"]
packages = [{ include = "blelocation" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
bleak = "^0.22.3"
ipykernel = "*"
joblib = "*"
scikit-learn = "*"
pandas = "*"
seaborn = "*"
japanize-matplotlib = "*"
setuptools = "*"
slack-sdk = "*"

[tool.poetry.dev-dependencies]
pandas-stubs = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
# Ignore missing imports for packages without stubs
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "japanize_matplotlib.*"
ignore_errors = true
